pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "setuptools-scm"]
build-backend = "setuptools.build_meta"

[project]
name = "agentpulse"
version = "0.1.0"
description = "Self-healing health monitor for AI Agents"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.11"
dependencies = [
    "click>=8.0",
    "httpx>=0.25",
    "rich>=13.0",
    "pyyaml>=6.0",
]

[project.scripts]
agentpulse = "agentpulse.cli:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
